requirements.txt
fastapi>=0.110
uvicorn>=0.30
pillow
pyyaml
jinja2
sqlalchemy>=2.0
alembic
psycopg2-binary
imagehash
# NSFW Detection dependencies (optional)
timm>=0.9.0
torch>=2.0.0